fix for win10
Didnt realise last update broke win 10. well not broke but threw an error related to the sort_banlist function. rewrote function and included windows . could not for the life of me get in-built functions is_windows or is_posix to work kept trying to find linux path so imported sys . now it launches properly on win 10 with no error.