Resolves Issue #58
The proper configuration variable to test is EMAIL_TIMER.  is_config_enabled() only tests for OFF or ON, and it it's not ON it's OFF.  So the numerical value in the config variable EMAIL_FREQUENCY was setting the boolean to false, which caused the routine not immediately send the email rather than queuing it up.  

The other part of this fix is to actually create the /var/artillery/src/program_junk directory.  If it does not exists then the temporary email log file cannot be created and written to.